Crate resource_proof

Resource proof

An mechanism to prove resource avaliability of a machine prior to it joining a network. This crate will provide the creation and confirmation algorithms. It is suggested that network nodes will require minimum resources to confirm the proof, but joining nodes will have a significantly higher resource requirementto attempt such a joining proof.

Github repository

Structs

ResourceProof
Holds the prover requirements

Functions

hash
Simple wrapper around tiny-keccak for use with deques