Struct resource_proof::ResourceProof

pub struct ResourceProof {
    // some fields omitted
}

Holds the prover requirements

Methods

impl ResourceProof

fn new(min_size: usize, difficulty: u8) -> ResourceProof

Rounds will factor how large the message to send is. It also has a slight impact on the hash power required, but this is minimal.

fn create_proof(&self, nonce: &[u8]) -> VecDeque<u8>

Use some data from them and some from you to create a proof.

fn validate_proof(&self, nonce: &[u8], proof: &VecDeque<u8>) -> bool

Use some data from them and some from you to confirm a proof.